+++ demo/app/tests/dom/ListViewNode.ts
@@ -1,10 +1,10 @@
-import { createElement, NativeElementNode } from 'svelte-native/dom'
+import { createElement, NativeViewElementNode } from 'svelte-native/dom'
 import { ListView } from 'tns-core-modules/ui/list-view'
 
 import ListViewMultiTemplate from './ListViewMultiTemplate.svelte'
 
 describe('ListViewNode', function () {
-    let test_subject: NativeElementNode
+    let test_subject: NativeViewElementNode<ListView>
     before(async function () {
         let el = createElement('fragment');
         let harness = new ListViewMultiTemplate({ target: el });
@@ -13,7 +13,7 @@ describe('ListViewNode', function () {
     })
 
     it('detects keyed templates', async function () {
-        let listview = test_subject.nativeView as ListView;
+        let listview = test_subject.nativeView;
         assert.equal(listview.itemTemplates.length, 2, `expected 2 keyed item templates`)
     })