+++ src/dom/index.ts
@@ -13,9 +13,9 @@ export { default as StyleElement } from './svelte/StyleElement'
 
 export { default as NativeElementNode, NativeElementPropConfig, NativeElementPropType } from './native/NativeElementNode'
 export { default as NativeViewElementNode } from './native/NativeViewElementNode'
+export { default as ActionBarElement } from './native/ActionBarElement'
 export { default as FrameElement } from "./native/FrameElement"
 export { default as TabsElement } from './native/TabsElement'
-export { default as TabStripElement } from './native/TabStripElement'
 export { default as PageElement } from './native/PageElement'
 export { default as ListViewElement, SvelteKeyedTemplate } from './native/ListViewElement'
 export { default as BottomNavigationElement } from './native/BottomNavigationElement'